// types/beatmap.ts
export interface Beatmapset extends BeatmapsetCompactBase {
	availability: BeatmapsetCompactAvailability;
	/** float */
	bpm: number;
	can_be_hyped: boolean;
	/**
	 * Username of the mapper at the time of beatmapset creation
	 */
	creator: string;
	discussion_enabled: boolean;
	discussion_locked: boolean;
	hype: null | BeatmapsetCompactHype;
	is_scoreable: boolean;
	last_updated: Timestamp;
	legacy_thread_url?: string;
	nominations_summary: BeatmapsetCompactNominationsSummary;
	/**
	 * See Rank status for list of possible values
	 */
	ranked?: RankedStatus;
	ranked_date?: Timestamp;
	source?: string;
	storyboard: boolean;
	submitted_date?: Timestamp;
	tags: string;

	has_favourited: boolean;
}

/**
 * Represent a beatmap.
 *
 * https://osu.ppy.sh/docs/index.html#beatmapcompact
 */
export interface BeatmapCompact {
	/** float */
	difficulty_rating: number;
	/** integer */
	id: number;
	mode: "osu" | "taiko" | "mania" | "fruits";
	/**
	 * See Rank status for list of possible values.
	 */
	status: string;
	total_length: number; // integer
	version: string;
	// Optional attributes:
	/**
	 * Beatmapset for Beatmap object, BeatmapsetCompact for
	 * BeatmapCompact object. null if the beatmap doesn't
	 * have associated beatmapset (e.g. deleted).
	 */
	beatmapset?: null | Beatmapset; // | BeatmapCompact
	checksum?: string;
	failtimes?: Failtimes;
	/** integer */
	max_combo?: number;
}

/**
 * Represent a beatmap.
 * This extends BeatmapCompact with additional attributes.
 *
 * https://osu.ppy.sh/docs/index.html#beatmap
 */
export interface Beatmap extends BeatmapCompact {
	/** float */
	accuracy: number;
	/** float */
	ar: number;
	/** integer */
	beatmapset_id: number;
	user_id: number;
	/** float */
	bpm: number;
	convert: boolean;
	/** integer */
	count_circles: number;
	/** integer */
	count_sliders: number;
	/** integer */
	count_spinners: number;
	/** float */
	cs: number;
	deleted_at?: Timestamp;
	/** float */
	drain: number;
	/** integer */
	hit_length: number;
	is_scoreable: boolean;
	last_updated: Timestamp;
	/** integer */
	mode_int: number;
	/** integer */
	passcount: number;
	/** integer */
	playcount: number;
	/**
	 * See Rank status for list of possible values.
	 *
	 * integer
	 */
	ranked: RankedStatus;
	url: string;
}